+++ changelog.md
@@ -31,6 +31,8 @@ easier and more straight-forward.
 * define-vars, datacontext, and componentdata blocks are now parsed once when HTML is loaded (not on demand)
 * components now fire 'mount' event internally (as well as externally)
 * change to grammar to allow functions (fn) to receive named parameters
+* added spaceship '<=>' operator for comparison
+* added new fn:uppercase and fn:lowercase string functions
 * bugfix: class.[class] was not being properly set to false when set to the Hibiki value false
 * bugfix: and/or operators were not correctly evaluating 'noattr' as false
 * bugfix: fix component defaults, define-vars, and h-withcontext, to never update mobx state